Prepare Drug Episode Data from Flatiron drug episodes data

This function loads, filters, and processes drug episode data for cancer patients, focusing on advanced or metastatic treatments. It cleans and merges treatment episodes, filters based on the specified treatment given at line 1, 2 or 3, and integrates progression, mortality and last contact data.

Usage

prep_episode_data(
  tumour = "nsclc",
  treatment = "cisplatin|carboplatin",
  drug_separator = ",",
  overlap_threshold = 1,
  exact = FALSE
)

Arguments

tumour

Character. Tumour type to load. Must match the drug episodes files in the Flatiron exploratory round folder. (e.g., "nsclc").

treatment

Character. Regex string of drug names to filter for (e.g., "cisplatin|carboplatin").

drug_separator

A character string separating multiple drugs in linename. Default is ",".

overlap_threshold

Integer: number of shared drugs required to collapse lines. Default = 1.

exact

boolean: to specify whether we want an exact match to the drug name or a combination therapy

Value

A cleaned and filtered data.frame of drug episodes with progression, death, and follow-up dates integrated

Examples

if (FALSE) { # \dontrun{
  episodes <- prep_episode_data(tumour = "nsclc", treatment = "cisplatin|carboplatin")
  head(episodes)
} # }