Substance Use Treatment Dataset

A simulated dataset containing information about substance use treatment clients, their visits, medication, outcomes, and demographics.

Usage

substance_use

Format

A tibble with 16,823 rows and 13 variables:

client_id

Integer ID for each client

visit_date

Date of clinical visit

substance

Factor with substance types (e.g., Alcohol, Opiates)

medication

Factor with medication types (e.g., Acamprosate, Naltrexone)

frequency_of_use

Ordered factor of use frequency (Monthly, Several times per week, Daily)

age

Numeric client age

gender

Factor of client gender

race

Factor of client race

housing

Ordered factor of housing status (Unhoused, Unstable, Stable)

mental_health_disorder

Logical indicating presence of mental health disorder

discontinued

Logical indicating if treatment was discontinued

died_ever

Logical indicating if client died at any point in study

died_last_visit

Logical indicating if client died after their last recorded visit

Source

Simulated data created to reflect realistic patterns observed in substance use treatment programs.

Examples

summary(substance_use)
#>    client_id      visit_date           substance    
#>  Min.   :   1   Min.   :2020-01-01   Alcohol:12611  
#>  1st Qu.:1199   1st Qu.:2020-01-01   Opiates: 4212  
#>  Median :2475   Median :2020-03-31                  
#>  Mean   :2466   Mean   :2020-09-05                  
#>  3rd Qu.:3707   3rd Qu.:2021-01-25                  
#>  Max.   :5000   Max.   :2024-12-05                  
#>                   medication                 frequency_of_use      age       
#>  None                  :2571   Occasional            :1716    Min.   :18.00  
#>  Disulfiram            :2788   Monthly               :1904    1st Qu.:30.00  
#>  Acamprosate           :4083   Weekly                :2871    Median :41.00  
#>  Naltrexone            :4502   Several times per week:4931    Mean   :41.32  
#>  Methadone             :1205   Daily                 :5401    3rd Qu.:53.00  
#>  Buprenorphine-Naloxone:1674                                  Max.   :65.00  
#>         gender           race            housing      mental_health_disorder
#>  Male      :7989   White   :10025   Stable   :12926   Mode :logical         
#>  Female    :8173   Black   : 2619   Temporary: 2028   FALSE:11153           
#>  Non-binary: 661   Hispanic: 2557   Unstable :  997   TRUE :5670            
#>                    Asian   :  866   Unhoused :  872                         
#>                    Other   :  756                                           
#>                                                                             
#>  discontinued    died_ever       died_last_visit
#>  Mode :logical   Mode :logical   Mode :logical  
#>  FALSE:13203     FALSE:13621     FALSE:15504    
#>  TRUE :3620      TRUE :3202      TRUE :1319     
#>                                                 
#>                                                 
#>